# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Snaptic API",
    "version": "1.0.0",
    "description": "API documentation for the Snaptic Attendance System.",
    "contact": {
      "name": "Tushar Sahu",
      "email": "sahutushar532@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in a cookie"
      }
    },
    "schemas": {
...
      "Class": {
        "type": "object",
        "required": [
          "name",
          "startDate",
          "endDate",
          "startTime",
          "daysOfWeek"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "09:00"
          },
          "duration": {
            "type": "number",
            "default": 60
          },
          "daysOfWeek": {
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 0,
              "maximum": 6
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "studentCount": {
            "type": "number"
          }
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/classes": {
      "get": {
        "summary": "Get all classes for the current user",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of classes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new class",
        "tags": [
          "Classes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Class created"
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "summary": "Get class details by ID",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class details"
          }
        }
      },
      "put": {
        "summary": "Update a class",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a class",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class deleted"
          }
        }
      }
    },
    "/classes/bulk/status": {
      "put": {
        "summary": "Bulk update class status",
        "tags": [
          "Classes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids",
                  "status"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "archived"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Classes updated"
          }
        }
      }
    },
    "/classes/bulk": {
      "delete": {
        "summary": "Bulk delete classes",
        "tags": [
          "Classes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Classes deleted"
          }
        }
      }
    },
    "/classes/{id}/students": {
      "post": {
        "summary": "Enroll a student in a class",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student enrolled"
          }
        }
      }
    },
    "/classes/{id}/enrollments/import": {
      "post": {
        "summary": "Import students from another class",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sourceClassId"
                ],
                "properties": {
                  "sourceClassId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Students imported"
          }
        }
      }
    },
    "/classes/{id}/students/{studentId}": {
      "delete": {
        "summary": "Remove a student from a class",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student removed"
          }
        }
      }
    },
    "/classes/students/search": {
      "get": {
        "summary": "Search for students by email or name",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of students found"
          }
        }
      }
    },
    "/reports/class/{classId}": {
      "get": {
        "summary": "Get a summary report for a specific class",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class report data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalSessions": {
                          "type": "number"
                        },
                        "attendancePercentage": {
                          "type": "number"
                        },
                        "totalPresent": {
                          "type": "number"
                        },
                        "totalAbsent": {
                          "type": "number"
                        }
                      }
                    },
                    "students": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "studentId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "presentCount": {
                            "type": "number"
                          },
                          "absentCount": {
                            "type": "number"
                          },
                          "percentage": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}